requirements.txt
torch
torchvision
pandas
matplotlib
pyyaml
azure-storage-blob
google-cloud-storage
tqdm
psutil